+++ src/components/index.ts
@@ -4,3 +4,5 @@ export * from './sidebar';
 export * from './dropdown';
 export * from './card';
 export * from './table';
+export * from './tabs';
+export * from './menu';