+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.48](https://github.com/getpingback/ui/compare/v0.0.47...v0.0.48) (2023-12-18)
+
 ### [0.0.50](https://github.com/getpingback/ui/compare/v0.0.49...v0.0.50) (2023-12-22)
 
 ### [0.0.49](https://github.com/getpingback/ui/compare/v0.0.43...v0.0.49) (2023-12-22)
@@ -16,27 +18,47 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Bug Fixes
 
+* close menu on back step ([7cc8a6f](https://github.com/getpingback/ui/commits/7cc8a6febe3072e1439bf1bf689688b119414450))
+
+### [0.0.47](https://github.com/getpingback/ui/compare/v0.0.46...v0.0.47) (2023-12-18)
+
 * **sidebar:** decrease animation time ([a811c65](https://github.com/getpingback/ui/commits/a811c65c0436db1d0256de3df15a39a7a57019c0))
 
 ### [0.0.47](https://github.com/getpingback/ui/compare/v0.0.43...v0.0.47) (2023-12-22)
 
 
+
 ### Bug Fixes
+
+* stop propagation ([0a42005](https://github.com/getpingback/ui/commits/0a42005b044e5f5c7e351b5780a06644492b5320))
+
+### [0.0.46](https://github.com/getpingback/ui/compare/v0.0.45...v0.0.46) (2023-12-15)
 
 * **sidebar:** decrease animation time ([a811c65](https://github.com/getpingback/ui/commits/a811c65c0436db1d0256de3df15a39a7a57019c0))
 
 ### [0.0.46](https://github.com/getpingback/ui/compare/v0.0.43...v0.0.46) (2023-12-22)
 
 
+
 ### Bug Fixes
+
+
+* close dropdown on click ([94282db](https://github.com/getpingback/ui/commits/94282db03136cf6a41e7484d1e86e8c40b452e3b))
+
+### [0.0.45](https://github.com/getpingback/ui/compare/v0.0.44...v0.0.45) (2023-12-15)
 
 * **sidebar:** decrease animation time ([a811c65](https://github.com/getpingback/ui/commits/a811c65c0436db1d0256de3df15a39a7a57019c0))
 
 ### [0.0.45](https://github.com/getpingback/ui/compare/v0.0.43...v0.0.45) (2023-12-22)
 
 
+
 ### Bug Fixes
 
+* dropdown structure ([782ed88](https://github.com/getpingback/ui/commits/782ed883f9ad5189a927cb3c3aa5d1edf48c6e2d))
+* dropdown tests ([7adb87a](https://github.com/getpingback/ui/commits/7adb87a98655e0af6f95e46a7a80b15a5994d6e7))
+
+### [0.0.44](https://github.com/getpingback/ui/compare/v0.0.43...v0.0.44) (2023-12-15)
 * **sidebar:** decrease animation time ([a811c65](https://github.com/getpingback/ui/commits/a811c65c0436db1d0256de3df15a39a7a57019c0))
 
 ### [0.0.44](https://github.com/getpingback/ui/compare/v0.0.43...v0.0.44) (2023-12-22)
@@ -44,6 +66,7 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Bug Fixes
 
+* dropdown and  menusub ([1459866](https://github.com/getpingback/ui/commits/14598664239c1bc343aa103f9c7b4a128907b226))
 * **sidebar:** decrease animation time ([a811c65](https://github.com/getpingback/ui/commits/a811c65c0436db1d0256de3df15a39a7a57019c0))
 
 ### [0.0.43](https://github.com/getpingback/ui/compare/v0.0.42...v0.0.43) (2023-12-12)